Avoid killing clients with sigpipe when transfering clipboard data.
When transfering selection data from sender to receiver, normally they
both block on the write and read. However blocking on the read when
retaining the data is sub-optimal for the compositor. Therefore it
uses non-blocking I/O (*) but this seems to introduce a chance of
having the clients terminating with SIGPIPE signal due to trying to
write when the compositor has already closed the handle and started to
retain another selection from another client. (all this caused by a
rapid succession of QClipboard::setMimeData calls)

The patch ensures that the read end of the pipe is not closed until
all data has been read, even if a newer selection gets activated
meanwhile, in which case the data (that became obsolete) is simply
ignored.

(*) The patch also enables real non-blocking I/O which was missing
for some reason, until now the compositor blocked when the first
reaady to read signal came.