Enable selections to be offered directly from the compositor.
The new function QWaylandCompositor::overrideSelection() allows
implementing clipboard functionality between the compositor and the
clients. This becomes important when the compositor process has UI
running in it because that implies offering and receiving selection
data not just between the clients but also between the compositor
(acting as a client in this respect) and clients.

Change-Id: I15575101bbef3edd71e23b7fe1f898e5803b37c6
Reviewed-by: Jørgen Lind <jorgen.lind@nokia.com>