Removing Wayland::SurfacePrivate
And moving Wayland::SurfaceBuffer into its own file.

Also remove the extra functions in Wayland::Surface which just relayed
down to extended surface. The pattern is that WaylandSurface is the
entry point so we check there is the Wayland::Surface has the extension
we'r looking for, then calling it directly from WaylandSurface.

Last,  the data of an extension is suppose to be members of the
extension, so I moved the properties to be members of ExtendedSurface

Change-Id: I5877671a17ee873b7644094f695fcd3e4677964a
Reviewed-by: Samuel Rødal <samuel.rodal@nokia.com>