Pass the correct window on to QWindowSystemInterface.
Sometimes, the window's coordinate system is not bounded by
the coordinates of the QScreen (such as if the window uses a
landscape window orientation on a portrait screen).
QGuiApplication::topLevelAt() will fail to find this window, but since
we have the window already at the wayland plugin level we just use that.

We already did this for mouse events, so it's natural to do it here as
well.

Change-Id: Id086dd940aa1cf4eb4480117d243ab8a76da6cca
Reviewed-by: Laszlo Agocs <laszlo.p.agocs@nokia.com>