Send expose from qwindow-compositor
Otherwise apps that (correctly) only start rendering on QEvent::Expose
will not render at all due to not receiving the event.

Turns out calling sendOnScreenVisibilityChange() is not that simple:
It does nothing until the extendedsurface instance is created and that
happens only after a roundtrip to the client. Therefore the compositor
listens to a new signal in WaylandSurface that indicates when the
extendedsurface-related functions become functional.

Change-Id: I9fdac4a98e2ad5d8c913c2a7aeef5b313f27c50a
Reviewed-by: Jørgen Lind <jorgen.lind@nokia.com>