Handle events coming from show() properly in qwindow-compositor
By not having the event handler set up properly before calling
show() we lose the expose events for platforms that send a sync
expose from setVisible (the default QPlatformWindow implementation).
On platforms like xcb this was not visible because the expose came
later, but with kms the initial expose was lost completely.

Change-Id: I16848be35f4d07df8053c66f7d39633a23e45df3
Reviewed-by: Jørgen Lind <jorgen.lind@nokia.com>