Set surface class name.
The class name identifies the general class of applications to
which the surface belongs. The class is the name of the desktop
file and is derived from the process name.

We might want to add a new method to QWindow in order to set
a custom class name.

It's quiet similar to WM_CLASS on X11, for more information
see https://live.gnome.org/GnomeShell/ApplicationBased

Change-Id: I5be90fb45e82e79710dda376f42c18c126d2a80f
Reviewed-by: Andy Nichols <andy.nichols@digia.com>