Fix the pointer cursor handling (QtCompositor side)
We now use the wayland-cursor library for handling pointer cursors.
This required some API changes in QtCompositor, as cursors are now
represented by wl_surfaces instead of just wl_buffer data.  This means
that when a new cursor is set by a client, you will receive a surface
which you can generate cursors from.  This Surface will be painted to
when the cursor is changed.  This also means that not all
WaylandSurfaces are window surfaces, so it's important that you check
that a WaylandSurface has a shell surface before treating it like a
window surface.  An API has been added for checking if there is a
shellSurface available in a WaylandSurface.

The qwindow compositor example demonstrates the correct usage of these
new API's.

Change-Id: Ife6b417f1048fd926a3499a139de07bceaebc727
Reviewed-by: Jørgen Lind <jorgen.lind@gmail.com>