Disable desktop settings awareness for now
Loading the KDE and Gnome settings for Wayland clients don't yet make
sense.  For the platform theme we check to see whether we are in a KDE
or Gnome environment, then set the default style hints based on that.
Right now for the Wayland platform, it makes more sense to use the
generic UNIX platform theme.

If there are ever KDE or Gnome Wayland environments available, it may be
worth reconsidering, but for now it just causes problems(like the GTK+
style being loaded even though it has X11 dependencies).

Change-Id: I13017c46f47ae34df884d2070a239c16ffbee850
Reviewed-by: Jørgen Lind <jorgen.lind@gmail.com>
Reviewed-by: J-P Nurmi <jpnurmi@digia.com>