Read and process Wayland events in separate thread
This commit introduces a new thread that is to be considered the main
Wayland thread.  This is the thread where we read from the Wayland
socket.  We are now using the Wayland event queue feature to make sure
that events are dispatched in the correct threads.

This commit also re-enables support for threaded OpenGL rendering.

Change-Id: I4f7bfdb6c008be2232bb98b63818a984eae44abf
Reviewed-by: Jørgen Lind <jorgen.lind@gmail.com>