WaylandSurfaceItem: add property resizeSurfaceToItem
Previously if a WaylandSurfaceItem's size changed, this only meant that
we were stretching the texture of the WaylandSurface it was
representing. When the resizeSurfaceToItem property is enabled, changing
the size of the WaylandSurfaceItem sends a configure event to the
Wayland ShellSurface for the client surface to be resized to match the
size of the WaylandSurfaceItem. This can be useful when the compositor
wants to enforce a size for clients.

Change-Id: Ie1bd387f6579323871fb3da542f0361e37aa1976
Reviewed-by: Andrew Knight <andrew.knight@digia.com>
Reviewed-by: Jørgen Lind <jorgen.lind@gmail.com>
Reviewed-by: Andy Nichols <andy.nichols@digia.com>