Fix window decorations to be thread safe
Rather than have window decorations that paint directly to the paint
device representing the window surface, window decorations are now
generated on demand into a backing store based on their current state,
and then are rendered to a surface when needed.

There is still a problem with resizing though that needs to be
addressed.

Change-Id: I7516628239eb0f9ae46d49bcb4bb2cb94ec6a617
Reviewed-by: Jørgen Lind <jorgen.lind@digia.com>