Improve the cursor handling
The decorations now set the cursor without using QGuiApplication's
override cursor API, but by setting directly the InputDevice's cursor.
This way it changes the cursor of the relevant seat only, and it also
fixes a bug where the cursor did not change back to the arrow after
moving a window.

Change-Id: I68e8cb8ff2367674dc376694a696baceca645843
Reviewed-by: Andy Nichols <andy.nichols@digia.com>