Integrate QWindow::Visibility into Compositor API
This enhances the onscreen_visibility extension to use QWindow's
visibility API. This way, the compositor can set the QWindow::visibility
on a Qt client. Any client that updates its window type through wl_shell
will also have its QWaylandSurface::visibility property updated, and the
compositor can choose to composite the surface accordingly. This also
simplifies exposure handling in QWaylandWindow so that it can use the
base implementation.

The existing use of this extension, setClientRenderingEnabled, is kept for
compatibility, and maps to QWindow::Hidden (false) and
QWindow::AutomaticVisibility (true).

Change-Id: If009649b98c792ca79585fbb09ea7912efeef7da
Reviewed-by: Andy Nichols <andy.nichols@digia.com>