Do not rely on deprecated wl_buffer struct
wl_buffer objects are not necessary deprecated wl_buffer structs
anymore. See for example the mesa commit:
http://cgit.freedesktop.org/mesa/mesa/commit/?id=8d29b5271a2e66fc78436be31ed6748ff006f0cb

Add a bufferSize() method to QWaylandGraphicsHardwareIntegration to
query the size of the buffer (instead of using width and height from the
wl_buffer struct).

Use wl_resource instead of wl_buffer as argument type in the
QWaylandGraphicsHardwareIntegration methods.

Change-Id: Id7ca7e84c2abe31f7448fe59c2d0005edcb4ec25
Reviewed-by: Andy Nichols <andy.nichols@digia.com>