Remove QWaylandSurface::setSize because it is not meaningful
Setting the surface size from the compositor is not meaningful; there
should be a roundtrip from the client before the size of the surface
is considered changed. This makes size a read-only property and makes
the requestSize (formerly sendConfigure) method invokable so that it can
be used from e.g. QML. The name change to requestSize follows the
convention that the size property will likely, but not necessarily,
change when a client-dependent property is set.

Change-Id: I4d2cfdb938a92f9d80b4fd1a443f0a659f73c690
Reviewed-by: Pier Luigi Fiorini <pierluigi.fiorini@gmail.com>
Reviewed-by: Andy Nichols <andy.nichols@digia.com>