Call sendRelease before unlockNativeBuffer
Unlock native buffer can call to destroy egl image.
If a wayland buffer is bound to the egl image and the buffer
resource is already destroyed by client, the buffer resouce
in compositor will be destroyed with destroying the egl image.
On sendRelease, m_buffer is null since destroy_listener_callback
is called by unlockNativeBuffer of HW integration.

Change-Id: I8fcaec12b4acab2c9d5fb2f2b4abc1e40654f014
Reviewed-by: Laszlo Agocs <laszlo.agocs@digia.com>
Reviewed-by: Andy Nichols <andy.nichols@digia.com>