Just exit instead of asking the application to quit when the connection goes down.
As discussed in abe1be628da687b1943a41e5411f34f646aeca86, the existing approach
is ineffective in some cases in that it merely requests that the application
shut down. In the case of misbehaving threads (for example), the application can
sit around indefinitely instead of quitting as it should.

Additionally, add a warning to make it clear what happened.

This also matches up with the xcb plugin's behavior.

Change-Id: Iabc17dc268c849635263a16754b9f3bfa7bd9e4a
Reviewed-by: Giulio Camuffo <giuliocamuffo@gmail.com>
Reviewed-by: Andrew Knight <andrew.knight@digia.com>
Reviewed-by: Laszlo Agocs <laszlo.agocs@digia.com>
Reviewed-by: Andy Nichols <andy.nichols@digia.com>