Add a preprocessor protocol name variable
We should convert - to _ in the preprocessor protocol name since - will
lead to a new token, but we don't since it is better to fail early here
and the official wayland-scanner doesn't do this either. The generated
c++ header will give a compiler error when being used which gives the
user a chance to change the protocol name.

Change-Id: I049bdecacc8ed9a70c0879494d4d358ce53f3320
Reviewed-by: Andy Nichols <andy.nichols@digia.com>