Fix crash when resizing a window in brcm-egl
This occurs for example when a Qt Quick2 application calls
showFullScreen() while a rendering is already in progress.
The GUI thread was calling destroyEglSurfaces() from setGeometry() and
thus destroying the buffers, while the render thread was trying to use
the buffers from swapBuffers().

There was a division by zero in swapBuffers() with the modulo m_count
operation, because EGL surfaces were destroyed and m_count was 0.

This fix is quite fragile, proper locking on buffers would be better.

Change-Id: I41fcb2fcebec84c0dfc82ce56c6e323d3722b30a
Reviewed-by: Laszlo Agocs <laszlo.agocs@theqtcompany.com>