Partially revert "Error out if dependencies aren't met."
This breaks building from qt.pro in qt5.git when doing integration for the whole
of Qt on CI nodes that don't have wayland installed.

This isn't an optimal solution of course, but it's pragmatic. Maybe it can be
revisited if/when per-module configures are ever a thing.

We reword the warnings to make it clear that QtWayland will not be built under
these circumstances.

This partially reverts commit 5b740c6246e092bc49b029bb7b12b614c4d0ce1e.

Change-Id: I07bd95a3a0349ad37a9025cdf5fd870c348d3799
Reviewed-by: Andrew Knight <andrew.knight@digia.com>