Put the pointers current view and current possition into a separate class
This also, if the current mouse coordinate ends on the boundary of the
surface, then subtract 0.01 from the mouse coordinate. This is to work
around Qt's mouse event propogation which is QRectF::contains which is
including the boundary, and Wayland which uses a strict contains. We
cant just discard events since Qt handles also enter/leave and clicking
state etc.

Change-Id: I919651232a5027b9fe0cb15b6838d3b1fee682df