Introduce the surface role concept
This is a concept originating initially from an implementation detail
in Weston which was later formalized in Wayland 1.7.
A wl_surface by itself will not be shown on screen, it must have a 'role'
attched for that to happen. A role can be a shell surface, a pointer surface, ...
Once a role is set on a wl_surface it cannot be changed, attempting to do so
should send a protocol error. It is however allowed to destroy and recreate
a new i.e. wl_shell_surface for the same wl_surface.

This patch introduces the role and role handler concepts.
A class such as a shell surface implementation should inherit from
SurfaceRoleHandler, and have a 'static SurfaceRole *role()' function,
which will be used to check against the role the surface has when
the role handler is set to it.
This changes introduces the new plumbing code but does not port
existing code to it. Eventually this new API should become public.

Change-Id: I43dc809b86051b7528ae6c39f796b9a96719ffd0
Reviewed-by: Pier Luigi Fiorini <pierluigi.fiorini@gmail.com>
Reviewed-by: Laszlo Agocs <laszlo.agocs@theqtcompany.com>