Keep Surface's visibility is in sync between client and compositor.
In change fa32466916f1e0fbd2be28bd48df2d9edfcd5e5f we moved the
visibility variable from QtWayland::ExtendedSurface to
QtWayland::Surface. But we didn't move the code in the ShellSurface
that kept this state in sync with the client. As a result
the client and compositor would get out of sync.

Change-Id: I54a7cb9f4b0669d3f3839ac8a29d8c42407195e2
Reviewed-by: Giulio Camuffo <giulio.camuffo@jollamobile.com>