Add mode support to QWaylandOutput
Outputs usually have more than one mode, add an API to support them.

When sizeFollowsWindow is true, modes are replaced by one with the
window size and refresh rate.  In that circumstance the mode changes
when the window is resized.

The sizeFollowsWindow property default value is no longer true.

The setGeometry() method is gone as it doesn't make sense now, the
setWidth() and setHeight() methods are now private slots to resize
the resolution as the window resizes (and sizeFollowsWindow is true).

Refresh rate is expressed in mHz rather than Hz just like the
Wayland protocol.

A compositor implementation may choose to add modes if it has access
to hardware information, it will call addMode() for each mode and
then invoke the setCurrentMode() method that sends the modes list
to the client. The preferred mode is indicated with a boolean
parameter to the addMode() method.

Change-Id: Iffed4784ccef695c276ebd800172957f4cff3324
Task-number: QTBUG-49814
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>
Reviewed-by: Johan Helsing <johan.helsing@qt.io>