update keymap right away when no key is currently pressed
Keymap update was only performed when a key was pressed. A key press right after a keymap update was pretty long.
Keymap may therefore be updated if no key is currently pressed. It avoids waiting for the next key press.
It must also be set when last key was released and do no need to wait for the next key press.

Change-Id: I9646a02f3112c8beb0f1cf90139e2b69af55a9b0
Reviewed-by: Giulio Camuffo <giulio.camuffo@jollamobile.com>