Make it possible for QWaylandSurfaceItem to children of arbitrary items
The main problem is that the position of the QWaylandSurfaceItem is
changed through the position of the views setPos function. However, it
might be that we don't want the surface position to map it directly.

Rename QWaylandSurfaceView::setPos to setRequestedPos
Then let the compositors choose what to do with this position

Change-Id: I43c74baad688d2f3e49ca091bbf436c718e81318