Move the texture ownership from the QWaylandSurface
to the respective QWaylandSurfaceView. For all GL applications this will have no
significant impact, but will cause a memory regression for shm surfaces
with multiple views.

This is done to simplify the creation and destruction of textures,
especially in multi-threaded environments.

Also the same patch removes the y_inverted property and replaces it with
a origin property. Its done in the same patch as a lot of the code was
overlapping.

Change-Id: I4bce50c614c9ac3ba0580e0560339476eac03433