Remove output from QWaylandSurface
It is the views that belong to outputs.

This leads to a couple of interesting cases with enter and leave events,
but the implementation says you will get enter events for the first view
of a surface entering an output, and for the last view being removed
from an output.

Also to throttle a surface, there has to be 1 surface which takes care
of this. This is because if multiple views throttle, then the client
might render to quick.

Change-Id: If4bba380fd4d7f506fd769606cbdea4ce58b908d