Remove automatic initialisation of extensions
It is the specific compositor application responsibillity to create the
specific extensions. Some of the extensions are picked up with the
QWaylandExtension::findIn(QWaylandExtensionContainer *) function to work
around missing events/apis. It should be a goal not to have these calles
in the base implementation of QWayland[Compositor|Surface|Input..].

Change-Id: Iacd576f2e8a79ca0165b858b1e8f44cddad48d93