Make the default wl_shell available from QML
This finaly ties together how to use QWaylandQuickItems with different
shells

It was required to decouple QWaylandView from the QWaylandQuickItem
since QML doesn't play to well with muliple inheritance. The
QWaylandQuickItem can be retrieved from the QWaylandView which is now
conveniently a QObject. Also the QWaylandQuickItem owns the
QWaylandView. This architecture also leaves room for creating a
QWaylandWidget :)

Change-Id: Ib8a00e6f17f0f1bfc3ff244753f021c76db22cb1