Make it possible to construct QWaylandOutput
This requires QWaylandOutput to be initialized. This will happend when
the QWaylandCompositor of the QWaylandOutputSpace that is set on the
QWaylandOutput is created. QWaylandCompositor will send or post a Polish
event to the QWaylandOutput which again will call
QWaylandOutput::initialize. This function will create the global that
will be put on the wl_displays registry

This makes it possible to Create WaylandOutputs in QML (see the pure-qml
example) and also gives a
better programming model in C++

Change-Id: Iaf907ae18a283678eda6d04390dc9790ae0bced8