Add a DiscardFontBuffers property to QWaylandView
This is an optimisation property which can be set on a view. If you have
two views for a surface, then the frontBuffer will be released first
when the last view has released the frontBuffer. With this option all
views with this option set to true will discard the frontBuffer when the
throtlingView discards its frontbuffer

Change-Id: Id628ec60830549ea930c1fff950501de2b073e48
Reviewed-by: Paul Olav Tvete <paul.tvete@theqtcompany.com>