Introduce QWaylandTextureBufferAttacher
Current surface implementation has two separate pointers
to a buffer attacher and it can only be set via the base
classes setter. When set externally the QWaylandQuickSurface
is unaware of the change and assumes certain members to be
present, thus leading to crashes.

QWaylandSurface assumes ownership of the attacher instance.

This commit also exposes the raw wayland buffer to allow more
flexible attachers.

Change-Id: If2677b779a035222f92c8dad78b27427c48f0851
Reviewed-by: Giulio Camuffo <giulio.camuffo@jollamobile.com>